more structure refactors
